+++ packages/twenty-server/src/engine/subscriptions/dtos/object-record-event.dto.ts
@@ -1,21 +1,6 @@
 import { Field, ObjectType } from '@nestjs/graphql';
 
-import GraphQLJSON from 'graphql-type-json';
-
-@ObjectType('ObjectRecordEventProperties')
-export class ObjectRecordEventPropertiesDTO {
-  @Field(() => [String], { nullable: true })
-  updatedFields?: string[];
-
-  @Field(() => GraphQLJSON, { nullable: true })
-  before?: object;
-
-  @Field(() => GraphQLJSON, { nullable: true })
-  after?: object;
-
-  @Field(() => GraphQLJSON, { nullable: true })
-  diff?: object;
-}
+import { ObjectRecordEventPropertiesDTO } from 'src/engine/subscriptions/dtos/object-record-event-properties.dto';
 
 @ObjectType('ObjectRecordEvent')
 export class ObjectRecordEventDTO {
